+++ src/Caliburn.Light.Core/Events/WeakEventSourceBase.cs
@@ -1,5 +1,7 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.CompilerServices;
 
 namespace Caliburn.Light;
 
@@ -12,7 +14,8 @@ public abstract class WeakEventSourceBase<TEventHandler>
 {
     private readonly object _lockObject = new object();
 
-    private WeakEventList? _list;
+    private List<WeakReference<TEventHandler>>? _list;
+    private ConditionalWeakTable<object, List<TEventHandler>>? _cwt;
 
     /// <summary>
     /// Adds the specified event handler.
@@ -24,10 +27,17 @@ public abstract class WeakEventSourceBase<TEventHandler>
 
         lock (_lockObject)
         {
-            if (_list is null)
-                _list = new WeakEventList();
+            var target = GetTarget(eventHandler, _lockObject);
 
-            _list.AddHandler(eventHandler);
+            _list = _list is null
+                ? new List<WeakReference<TEventHandler>>(1)
+                : new List<WeakReference<TEventHandler>>(_list);
+
+            _list.RemoveAll(static wr => !wr.TryGetTarget(out var _));
+            _list.Add(new WeakReference<TEventHandler>(eventHandler));
+
+            _cwt ??= new ConditionalWeakTable<object, List<TEventHandler>>();
+            _cwt.GetOrCreateValue(target).Add(eventHandler);
         }
     }
 
@@ -41,7 +51,28 @@ public abstract class WeakEventSourceBase<TEventHandler>
 
         lock (_lockObject)
         {
-            _list?.RemoveHandler(eventHandler);
+            if (_list is null || _cwt is null) return;
+
+            _list = new List<WeakReference<TEventHandler>>(_list);
+            _list.RemoveAll(static wr => !wr.TryGetTarget(out var _));
+
+            var index = _list.FindIndex(wr => wr.TryGetTarget(out var handler) && Equals(handler, eventHandler));
+            if (index >= 0) _list.RemoveAt(index);
+
+            var target = GetTarget(eventHandler, _lockObject);
+
+            if (_cwt.TryGetValue(target, out var value))
+            {
+                value.Remove(eventHandler);
+                if (value.Count == 0)
+                    _cwt.Remove(target);
+            }
+
+            if (_list.Count == 0)
+            {
+                _list = null;
+                _cwt = null;
+            }
         }
     }
 
@@ -49,13 +80,33 @@ public abstract class WeakEventSourceBase<TEventHandler>
     /// Gets the handlers to raise the event.
     /// </summary>
     /// <returns>The list of active event handlers.</returns>
-    protected IReadOnlyList<TEventHandler> GetHandlers()
+    protected IEnumerable<TEventHandler> GetHandlers()
     {
         lock (_lockObject)
         {
             return _list is null
-                ? Array.Empty<TEventHandler>()
-                : _list.GetHandlers<TEventHandler>();
+                ? Enumerable.Empty<TEventHandler>()
+                : GetHandlers(_list);
+        }
+    }
+
+    private static object GetTarget(TEventHandler handler, object staticTarget)
+    {
+        var target = handler.Target ?? staticTarget;
+
+        // protect weak event handlers from being collected
+        if (target is IWeakEventHandler)
+            target = staticTarget;
+
+        return target;
+    }
+
+    private static IEnumerable<TEventHandler> GetHandlers(List<WeakReference<TEventHandler>> list)
+    {
+        for (var i = 0; i < list.Count; i++)
+        {
+            if (list[i].TryGetTarget(out var handler))
+                yield return handler;
         }
     }
 }